tools/nvmeblk.bt: force nvme_core module BTF to be loaded
bpftrace list_modules() does not yet support loading BTFs for
modules on ProbeType::kprobe and it is not clear if it will. To load
the BTF for modules so we can get structures for it which might be
private (and could conflict with other modules own structures) one
has to use a probe type which does load the BTF for the corresponding
module.

Since we already are adding a kprobe on nvme_setup_cmd() just use
the kretfunc probe on the same routine and have it do nothing.

This fixes the use case of nvmeblk.bt when we nvme_core is a module.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>